ai_agents/requirements.txt
# AI Agents Dependencies

# AI/ML Core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Agent Framework (optional - comment out if not needed)
# crewai>=0.80.0

# AI/ML Libraries
google-generativeai>=0.3.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Email Handling
# For Gmail: use App Password, not regular password
# No additional packages needed - using built-in imaplib and smtplib

# Scheduling
schedule>=1.2.0

# Data Processing (using prebuilt wheels)
pandas>=2.0.0
numpy>=1.26.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
